# contracts/openapi/portfolio-api.json
{
  "components": {
    "schemas": {
...
      "ChartResponse": {
        "properties": {
          "bars": {
            "items": {
              "$ref": "#/components/schemas/ChartBarResponse"
            },
            "type": "array"
          },
          "dataAsOf": {
            "format": "date-time",
            "type": "string"
          },
          "earningsMarkers": {
            "items": {
              "$ref": "#/components/schemas/EventMarkerResponse"
            },
            "type": "array"
          },
          "entryMarkers": {
            "items": {
              "$ref": "#/components/schemas/ChartMarkerResponse"
            },
            "type": "array"
          },
          "quality": {
            "type": "string"
          },
          "stopSeries": {
            "items": {
              "$ref": "#/components/schemas/StopSeriesResponse"
            },
            "type": "array"
          },
          "tradeMarkers": {
            "items": {
              "$ref": "#/components/schemas/EventMarkerResponse"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "ConfirmationRequest": {
        "properties": {
          "accountMappings": {
            "items": {
              "$ref": "#/components/schemas/AccountMappingRequest"
            },
            "type": "array"
          },
          "cashSetup": {
            "$ref": "#/components/schemas/CashSetupRequest"
          },
          "expectedVersion": {
            "format": "int64",
            "type": "integer"
          },
          "rowOverrides": {
            "items": {
              "$ref": "#/components/schemas/RowOverrideRequest"
            },
            "type": "array"
          }
        },
        "required": ["cashSetup"],
        "type": "object"
      },
      "ConfirmationResponse": {
        "properties": {
          "analysisRunId": {
            "format": "uuid",
            "type": "string"
          },
          "analysisState": {
            "type": "string"
          },
          "batchId": {
            "format": "uuid",
            "type": "string"
          },
          "cashflowReconciliation": {
            "$ref": "#/components/schemas/CashflowReconciliationResponse"
          },
          "cashRowCount": {
            "format": "int32",
            "type": "integer"
          },
          "closedPositionCount": {
            "format": "int32",
            "type": "integer"
          },
          "compensationRowCount": {
            "format": "int32",
            "type": "integer"
          },
          "idempotentReplay": {
            "type": "boolean"
          },
          "openPositionCount": {
            "format": "int32",
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "DipPreviewRequest": {
        "properties": {
          "aboveEma20": {
            "type": "boolean"
          },
          "breadthImproving": {
            "type": "boolean"
          },
          "breadthOversoldScore": {
            "format": "double",
            "type": "number"
          },
          "breakout5Day": {
            "type": "boolean"
          },
          "completeData": {
            "type": "boolean"
          },
          "completedTranches": {
            "format": "int32",
            "maximum": 4,
            "minimum": 0,
            "type": "integer"
          },
          "creditStable": {
            "type": "boolean"
          },
          "creditStressScore": {
            "format": "double",
            "type": "number"
          },
          "drawdownScore": {
            "format": "double",
            "type": "number"
          },
          "emergencyCashProtected": {
            "type": "boolean"
          },
          "marketDriven": {
            "type": "boolean"
          },
          "portfolioDrawdown": {
            "minimum": 0,
            "type": "number"
          },
          "rsiCross40": {
            "type": "boolean"
          },
          "tradingDaysSinceLastTranche": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "trendContextScore": {
            "format": "double",
            "type": "number"
          },
          "vixFalling": {
            "type": "boolean"
          },
          "vixPercentileScore": {
            "format": "double",
            "type": "number"
          },
          "volTermScore": {
            "format": "double",
            "type": "number"
          }
        },
        "required": ["portfolioDrawdown"],
        "type": "object"
      },
      "DipPreviewResponse": {
        "properties": {
          "action": {
            "type": "string"
          },
          "debug": {
            "type": "boolean"
          },
          "executionSubmitted": {
            "type": "boolean"
          },
          "reserveFraction": {
            "type": "string"
          },
          "ruleIds": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "setupScore": {
            "format": "double",
            "type": "number"
          },
          "trancheNumber": {
            "format": "int32",
            "type": "integer"
          },
          "triggerCount": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "EarningsReviewRequest": {
        "properties": {
          "classification": {
            "minLength": 1,
            "type": "string"
          },
          "eventCount": {
            "format": "int32",
            "maximum": 12,
            "minimum": 0,
            "type": "integer"
          },
          "overRiskLimit": {
            "type": "boolean"
          },
          "profitCushionR": {
            "type": "number"
          }
        },
        "required": ["classification", "profitCushionR"],
        "type": "object"
      },
      "EarningsReviewResponse": {
        "properties": {
          "action": {
            "type": "string"
          },
          "ruleIds": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "IntelligenceResponse": {
        "properties": {
          "earnings": {
            "$ref": "#/components/schemas/EarningsResponse"
          },
          "journal": {
            "items": {
              "$ref": "#/components/schemas/JournalResponse"
            },
            "type": "array"
          },
          "stop": {
            "$ref": "#/components/schemas/StopSnapshotResponse"
          },
          "thesis": {
            "$ref": "#/components/schemas/ThesisResponse"
          },
          "valuation": {
            "$ref": "#/components/schemas/ValuationResponse"
          }
        },
        "type": "object"
      },
...
      "PositionResponse": {
        "properties": {
          "accountId": {
            "format": "uuid",
            "type": "string"
          },
          "averageCost": {
            "type": "string"
          },
          "bucket": {
            "type": "string"
          },
          "classification": {
            "type": "string"
          },
          "classificationConfirmed": {
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "marketValue": {
            "type": "string"
          },
          "quantity": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "symbol": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "PreviewResponse": {
        "properties": {
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/ImportAccount"
            },
            "type": "array"
          },
          "batchId": {
            "format": "uuid",
            "type": "string"
          },
          "cash": {
            "items": {
              "$ref": "#/components/schemas/CashResponse"
            },
            "type": "array"
          },
          "dataAsOf": {
            "format": "date-time",
            "type": "string"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "holdings": {
            "items": {
              "$ref": "#/components/schemas/HoldingResponse"
            },
            "type": "array"
          },
          "status": {
            "type": "string"
          },
          "summary": {
            "$ref": "#/components/schemas/SummaryResponse"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Private, manual-execution-only portfolio decision support API",
    "title": "Portfolio Engine API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/etf-dip/preview": {
      "post": {
        "operationId": "preview_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DipPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DipPreviewResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["dip-cashflow-controller"]
      }
    },
...
    "/api/v1/portfolio-imports/{batchId}/confirm": {
      "post": {
        "operationId": "confirm",
        "parameters": [
          {
            "in": "path",
            "name": "batchId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmationResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["portfolio-import-controller"]
      }
    },
    "/api/v1/portfolio-imports/fidelity/preview": {
      "post": {
        "operationId": "preview_1",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": ["file"],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["portfolio-import-controller"]
      }
    },
...
    "/api/v1/positions/{id}": {
      "get": {
        "operationId": "position",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["portfolio-controller"]
      }
    },
...
    "/api/v1/positions/{positionId}/chart": {
      "get": {
        "operationId": "chart",
        "parameters": [
          {
            "in": "path",
            "name": "positionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "range",
            "required": false,
            "schema": {
              "default": "1Y",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChartResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["position-intelligence-controller"]
      }
    },
    "/api/v1/positions/{positionId}/earnings/review": {
      "post": {
        "operationId": "reviewEarnings",
        "parameters": [
          {
            "in": "path",
            "name": "positionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EarningsReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EarningsReviewResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["position-intelligence-controller"]
      }
    },
    "/api/v1/positions/{positionId}/intelligence": {
      "get": {
        "operationId": "intelligence",
        "parameters": [
          {
            "in": "path",
            "name": "positionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/IntelligenceResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["position-intelligence-controller"]
      }
    },
    "/api/v1/positions/{positionId}/journal": {
      "get": {
        "operationId": "journal",
        "parameters": [
          {
            "in": "path",
            "name": "positionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JournalResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["position-intelligence-controller"]
      }
    },
...
    "/api/v1/positions/{positionId}/stops/preview": {
      "post": {
        "operationId": "previewStop",
        "parameters": [
          {
            "in": "path",
            "name": "positionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StopPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StopPreviewResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["position-intelligence-controller"]
      }
    },
    "/api/v1/positions/{positionId}/thesis/confirm": {
      "post": {
        "operationId": "confirmThesis",
        "parameters": [
          {
            "in": "path",
            "name": "positionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmThesisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ThesisResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["position-intelligence-controller"]
      }
    },
...
    "/api/v1/trade-plans/preview": {
      "post": {
        "operationId": "preview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TradePlanPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TradePlanPreviewResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["portfolio-controller"]
      }
    },
...
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost"
    }
  ]
}